--- a/crypto/key.go
+++ b/crypto/key.go
@@ -282,7 +282,20 @@ func UnmarshalPublicKey(data []byte) (PubKey, error) {
 		return nil, ErrBadKeyType
 	}
 
-	return um(pmes.GetData())
+	pk, err := um(pmes.GetData())
+	if err != nil {
+		return nil, err
+	}
+
+	switch tpk := pk.(type) {
+	case *RsaPublicKey:
+		bcopy := make([]byte, len(data))
+		copy(bcopy, data)
+
+		tpk.cached = bcopy
+	}
+
+	return pk, nil
 }
 
 // MarshalPublicKey converts a public key object into a protobuf serialized