This is making GC realllly hot on the machines i've been watching.

Here's the before and after benchmark results:

Before:
```
why@WhyNet ~/g/s/g/l/g/crypto> go test -run XXXX -bench=BenchmarkRSAKey
goos: darwin
goarch: amd64
pkg: github.com/libp2p/go-libp2p-core/crypto
BenchmarkRSAKeyEqualsCheck-4   	  100000	     13415 ns/op	    5443 B/op	     118 allocs/op
PASS
ok  	github.com/libp2p/go-libp2p-core/crypto	4.706s

why@WhyNet ~/g/s/g/l/g/crypto> go test -run XXXX -bench=BenchmarkRSAKey -tags=openssl
# github.com/libp2p/go-libp2p-core/crypto.test
ld: warning: directory not found for option '-L/usr/local/opt/openssl@1.1/lib'
goos: darwin
goarch: amd64
pkg: github.com/libp2p/go-libp2p-core/crypto
BenchmarkRSAKeyEqualsCheck-4   	  100000	     14560 ns/op	    5024 B/op	      18 allocs/op
PASS
ok  	github.com/libp2p/go-libp2p-core/crypto	2.789s
```

After:

```
why@WhyNet ~/g/s/g/l/g/crypto> go test -run XXXX -bench=BenchmarkRSAKey
goos: darwin
goarch: amd64
pkg: github.com/libp2p/go-libp2p-core/crypto
BenchmarkRSAKeyEqualsCheck-4   	20000000	        69.6 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/libp2p/go-libp2p-core/crypto	6.666s

why@WhyNet ~/g/s/g/l/g/crypto> go test -run XXXX -bench=BenchmarkRSAKey -tags=openssl
# github.com/libp2p/go-libp2p-core/crypto.test
ld: warning: directory not found for option '-L/usr/local/opt/openssl@1.1/lib'
goos: darwin
goarch: amd64
pkg: github.com/libp2p/go-libp2p-core/crypto
BenchmarkRSAKeyEqualsCheck-4   	30000000	        52.5 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/libp2p/go-libp2p-core/crypto	3.788s
```