@Stebalien 2019-06-28
https://github.com/libp2p/go-libp2p/pull/665